refactor(vite-plugin): rename vite package to vite-plugin
