+++ packages/react-utils/package.json
@@ -25,7 +25,6 @@
 		"build:js:prod": "vite build --mode production",
 		"ae:build": "api-extractor run --local",
 		"ae:test": "api-extractor run"
-
 	},
 	"peerDependencies": {
 		"react": "^17 || ^18"
