+++ packages/vite-plugin/src/tsconfig.json
