+++ packages/vite-plugin/tsconfig.json
